requirements.txt
llama-index==0.10.1
llama-index-core==0.10.1
qdrant-client==1.7.0
aiobotocore==2.19.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aioitertools==0.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
boto3==1.36.18
botocore==1.36.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
courlan==1.3.2
dataclasses-json==0.6.7
dateparser==1.2.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
htmldate==1.9.3
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jusText==3.0.1
llama-cloud==0.1.12
llama-cloud-services==0.6.0
llama-index-agent-openai==0.4.5
llama-index-cli==0.4.0
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.4
llama-index-llms-openai==0.3.18
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-index-vector-stores-qdrant==0.4.3
llama-parse==0.6.0
lxml==5.3.1
lxml_html_clean==0.4.1
marshmallow==3.26.1
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.2
openai==1.61.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
portalocker==2.10.1
propcache==0.2.1
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3fs==2025.2.0
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
striprtf==0.0.26
tenacity==9.0.0
tiktoken==0.8.0
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
wrapt==1.17.2
yarl==1.18.3
coloredlogs==15.0.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastembed==0.5.1
filelock==3.17.0
flatbuffers==25.2.10
httptools==0.6.4
huggingface-hub==0.28.1
humanfriendly==10.0
Jinja2==3.1.5
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
onnxruntime==1.20.1
py_rust_stemmers==0.1.3
Pygments==2.19.1
python-multipart==0.0.20
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
starlette==0.45.3
sympy==1.13.3
tavily-python==0.5.1
tokenizers==0.21.0
typer==0.15.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2